+++ src/pages/message/editor.css
@@ -1,0 +1,13 @@
+@import 'tailwindcss';
+
+.ql-toolbar.ql-snow {
+  @apply bg-[#eeeeee] rounded-t-lg;
+}
+
+.ql-container.ql-snow {
+  @apply h-[250px] rounded-b-lg;
+}
+
+.ql-editor {
+  @apply text-base;
+}
